add update player when end match

+++ ChineseChessApi/src/main/java/com/data/dto/MatchDTO.java
@@ -25,6 +25,6 @@ public class MatchDTO extends AuditingDTO{
 
     private PlayerProfileDTO player2ProfileDTO;
 
-    private Long result;
+    private Integer result;
     
 }
